CookieTest: split `domain_matches()` specific tests to dedicated test class
