+++ tests/Cookie/CookieTest.php
@@ -125,56 +125,6 @@ final class CookieTest extends TestCase {
 		$this->assertSame('testvalue2', $data['requests-testcookie2']);
 	}
 
-	public function domainMatchProvider() {
-		return [
-			'Invalid check domain (type): null'         => ['example.com', null, false, false],
-			'Invalid check domain (type): boolean true' => ['example.com', true, false, false],
-			['example.com', 'example.com', true, true],
-			['example.com', 'www.example.com', false, true],
-			['example.com', 'example.net', false, false],
-
-			// Leading period
-			['.example.com', 'example.com', true, true],
-			['.example.com', 'www.example.com', false, true],
-			['.example.com', 'example.net', false, false],
-
-			// Prefix, but not subdomain
-			['example.com', 'notexample.com', false, false],
-			['example.com', 'notexample.net', false, false],
-
-			// Reject IP address prefixes
-			['127.0.0.1', '127.0.0.1', true, true],
-			['127.0.0.1', 'abc.127.0.0.1', false, false],
-			['127.0.0.1', 'example.com', false, false],
-
-			// Check that we're checking the actual length
-			['127.com', 'test.127.com', false, true],
-		];
-	}
-
-	/**
-	 * @dataProvider domainMatchProvider
-	 */
-	public function testDomainExactMatch($original, $check, $matches, $domain_matches) {
-		$attributes           = new CaseInsensitiveDictionary();
-		$attributes['domain'] = $original;
-		$cookie               = new Cookie('requests-testcookie', 'testvalue', $attributes);
-		$this->assertSame($matches, $cookie->domain_matches($check));
-	}
-
-	/**
-	 * @dataProvider domainMatchProvider
-	 */
-	public function testDomainMatch($original, $check, $matches, $domain_matches) {
-		$attributes           = new CaseInsensitiveDictionary();
-		$attributes['domain'] = $original;
-		$flags                = [
-			'host-only' => false,
-		];
-		$cookie               = new Cookie('requests-testcookie', 'testvalue', $attributes, $flags);
-		$this->assertSame($domain_matches, $cookie->domain_matches($check));
-	}
-
 	public function pathMatchProvider() {
 		return [
 			'Invalid check path (type): null'    => ['/', null, true],
